+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useCaseBlocksLogic/utils/useTabsToBlocksMap.tsx
@@ -21,7 +21,10 @@ import { z } from 'zod';
 import { handleZodError } from '@/common/utils/handle-zod-error/handle-zod-error';
 import { toast } from 'sonner';
 import { t } from 'i18next';
-import { useEditCollectionFlow } from '@/pages/Entity/components/Case/components/CaseOptions/hooks/useEditCollectionFlow';
+import {
+  EDIT_TEMPLATES,
+  useEditCollectionFlow,
+} from '@/pages/Entity/components/Case/components/CaseOptions/hooks/useEditCollectionFlow';
 
 export type TCaseBlocksCreationProps = {
   workflow: TWorkflowById;
@@ -145,7 +148,9 @@ export const useTabsToBlocksMap = ({
   const { mutate: mutateInitiateIndividualVerificationAndSendEmail } =
     useInitiateIndividualVerificationAndSendEmailMutation();
 
-  const { onEditCollectionFlow } = useEditCollectionFlow();
+  const { onEditCollectionFlow: onEditCompanyOwnership } = useEditCollectionFlow(
+    EDIT_TEMPLATES.COMPANY_OWNERSHIP,
+  );
   const getInitiateKycEvent = (nextEvents: string[]) => {
     if (nextEvents?.includes('start')) {
       return 'start';
@@ -291,7 +296,7 @@ export const useTabsToBlocksMap = ({
               ids,
               workflowId: childWorkflow?.id,
             }),
-        onEdit: onEditCollectionFlow({ steps: ['company_ownership'] }),
+        onEdit: onEditCompanyOwnership,
         reasons:
           childWorkflow?.workflowDefinition?.contextSchema?.schema?.properties?.documents?.items?.properties?.decision?.properties?.revisionReason?.anyOf?.find(
             ({ enum: enum_ }) => !!enum_,
@@ -315,6 +320,8 @@ export const useTabsToBlocksMap = ({
       } satisfies Parameters<typeof createKycBlocks>[0][number];
     },
     [
+      getStatusFromCheckStatus,
+      mutateInitiateIndividualVerificationAndSendEmail,
       getStatusFromTags,
       getInitiateKycEvent,
       getInitiateSanctionsScreeningEvent,
@@ -328,6 +335,7 @@ export const useTabsToBlocksMap = ({
       workflow?.workflowDefinition?.config?.isInitiateSanctionsScreeningEnabled,
       workflow?.workflowDefinition?.config?.isInitiateKycEnabled,
       workflow?.workflowDefinition?.config?.isKycEndUserEditEnabled,
+      onEditCompanyOwnership,
     ],
   );
   const directorToIndividualAdapter = useCallback(
@@ -390,7 +398,7 @@ export const useTabsToBlocksMap = ({
         onReuploadNeeded:
           ({ reason, ids }: { reason: string; ids: string[] }) =>
           () => {},
-        onEdit: onEditCollectionFlow({ steps: ['company_ownership'] }),
+        onEdit: onEditCompanyOwnership,
         reasons: [],
         isReuploadNeededDisabled: true,
         isApproveDisabled: true,
@@ -407,6 +415,9 @@ export const useTabsToBlocksMap = ({
       } satisfies Parameters<typeof createKycBlocks>[0][number];
     },
     [
+      mutateInitiateIndividualVerificationAndSendEmail,
+      onEditCompanyOwnership,
+      workflow?.workflowDefinition?.config?.language,
       workflow?.id,
       caseState.actionButtonsEnabled,
       workflow?.workflowDefinition?.config?.isInitiateKycEnabled,
@@ -448,7 +459,7 @@ export const useTabsToBlocksMap = ({
             documents: [...(director?.documents ?? []), ...(workflow?.context?.kycDocuments ?? [])],
           });
         }) ?? [],
-    [workflow, endUsers, directorToIndividualAdapter],
+    [workflow, endUsers, directorToIndividualAdapter, getStatusFromCheckStatus],
   );
 
   const personOfInterestToIndividualAdapter = useCallback(
@@ -513,7 +524,7 @@ export const useTabsToBlocksMap = ({
         onReuploadNeeded:
           ({ reason, ids }: { reason: string; ids: string[] }) =>
           () => {},
-        onEdit: onEditCollectionFlow({ steps: ['company_ownership'] }),
+        onEdit: onEditCompanyOwnership,
         reasons: [],
         isReuploadNeededDisabled: true,
         isApproveDisabled: true,
@@ -534,6 +545,11 @@ export const useTabsToBlocksMap = ({
       caseState.actionButtonsEnabled,
       workflow?.workflowDefinition?.config?.isInitiateKycEnabled,
       workflow?.workflowDefinition?.config?.isKycEndUserEditEnabled,
+      endUsers,
+      getStatusFromCheckStatus,
+      mutateInitiateIndividualVerificationAndSendEmail,
+      onEditCompanyOwnership,
+      workflow?.workflowDefinition?.config?.language,
     ],
   );