+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/EntityFieldGroup/components/EntityFields/helpers/build-entity-for-update.ts
@@ -2,7 +2,6 @@ import { IFormElement } from '@/components/organisms/Form/DynamicForm/types';
 import { IEntityFieldGroupParams } from '../../../EntityFieldGroup';
 
 import { AnyObject } from '@/common';
-import { TEntityFieldGroupType } from '../../../EntityFieldGroup';
 import { IEntity } from '../../../types';
 import { transform } from '../utils/transform';
 
@@ -10,14 +9,16 @@ export const buildEntityUpdatePayload = async (
   element: IFormElement<any, IEntityFieldGroupParams>,
   entity: IEntity,
   context: AnyObject,
-): Promise<{ entity: IEntity; entityType: TEntityFieldGroupType; ballerineEntityId?: string }> => {
+): Promise<{ entity: IEntity; ballerineEntityId?: string }> => {
   const entityToCreate = element.params?.httpParams?.createEntity?.transform
     ? await transform(context, entity, element.params!.httpParams?.createEntity.transform)
     : entity;
 
   return {
-    entity: entityToCreate,
-    entityType: element.params?.type as TEntityFieldGroupType,
+    entity: {
+      ...entityToCreate,
+      variant: entityToCreate.variant || element.params?.type,
+    },
     ballerineEntityId: entity.ballerineEntityId,
   };
 };