+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,26 @@
 # web-ui-sdk
 
+## 1.5.119
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.118
+
+## 1.5.118
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.117
+
+## 1.5.117
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.116
+
 ## 1.5.116
 
 ### Patch Changes