+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.142
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.118
+  - @ballerine/workflow-core@0.6.142
+
+## 0.6.141
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.117
+  - @ballerine/workflow-core@0.6.141
+
+## 0.6.140
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.116
+  - @ballerine/workflow-core@0.6.140
+
 ## 0.6.139
 
 ### Patch Changes